+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4
+Create a proper example file and updated the README
+Added private constructors for StlReader and StlWriter since all public APIs are static
+
 ## 0.0.3
 Added API docs for all public methods. Cleaned up some obsolete code.